Update And Add New Function To UserSongs
The new function getMetadataByUri is pretty self-explanatory, you give
it a Uri, it returns the Uri's song's metadata. A commentary explaining
how it works will be added in the future.

Also changed the old map-based returns from all the functions to now
return a SongMetadata data class with the same old values.